[msvc] Fix warnings, mainly casting to smaller types (#270)
If the checks by GitHub Actions checks are met, then it seems good to go.  Thanks for supporting FOSS4G!

+++ src/mbgl/renderer/layers/render_background_layer.cpp
@@ -72,7 +72,7 @@ void RenderBackgroundLayer::render(PaintParameters& parameters) {
             program.computeAllUniformValues(std::forward<decltype(uniformValues)>(uniformValues),
                                             paintAttributeData,
                                             properties,
-                                            parameters.state.getZoom());
+                                            static_cast<float>(parameters.state.getZoom()));
         const auto allAttributeBindings = program.computeAllAttributeBindings(
             *parameters.staticData.tileVertexBuffer,
             paintAttributeData,
